such a large refactor omg - removed loading of .sm and .osu - began groundwork for conversion of .sm and .osu to .yav - added loading and caching for .yav - restructured chart loading into individual difficulties rather than one song and multiple difficulties - organised lots of code into folders better - changed data types for some technical map storage to ~halve ram usage of note data - rewrote sprite loading system to support live skin swapping a bit better - rewrote spritebatch to be a bit more flexible and stuff - patched a ton of bugs and became aware of even more bugs than i've patched - started work on command line interface for certain tools i.e chart conversions

+++ Options/Colorizer.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Drawing;
 using YAVSRG.Gameplay;
-using YAVSRG.Beatmap;
+using YAVSRG.Charts.YAVSRG;
 
 namespace YAVSRG.Options
 {
@@ -102,7 +102,7 @@ namespace YAVSRG.Options
             foreach (GameplaySnap s in c.Notes.Points)
             {
                 s.colors = new int[c.Keys];
-                count = new BinarySwitcher(s.taps.value | s.holds.value).Count; //count number of lns/notes
+                count = new BinarySwitcher((ushort)(s.taps.value | s.holds.value)).Count; //count number of lns/notes
                 for (int i = 0; i < c.Keys; i++)
                 {
                     s.colors[i] = cs.GetColorIndex(count,c.Keys); //color notes in row based on number of notes. chord coloring.
